ops/jaeger/clients/python/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
googleapis-common-protos==1.63.0
grpcio==1.62.1
idna==3.7
importlib-metadata==7.0.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
protobuf==4.25.3
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
zipp==3.18.1
